backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
pandas
scikit-learn
python-multipart
PyPDF2
openpyxl
python-docx
pytesseract
Pillow
